Organise file locations
