+++ pyproject.toml
@@ -12,17 +12,13 @@ dependencies = [
     "ruff>=0.11.4",
     "streamlit>=1.44.1",
     "uvicorn>=0.34.0",
-    "pydantic-ai-slim[mcp]>=0.0.55",
+    "pydantic-ai-slim[]>=0.0.55",
     "mcp>=1.6.0",
     "numpy>=2.1.1",
     "opencv-python>=4.11.0.86",
     "pillow>=11.2.1",
     "ultralytics>=8.3.115",
 ]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
 
 [tool.uv]
 dev-dependencies = [
@@ -38,6 +34,3 @@ ignore = ["E501"]
 
 [tool.pytest.ini_options]
 addopts = '-q'
-
-[project.scripts]
-hello = "uv_docker_example:hello"
