+++ src/yolo/mcp-config.json
