+++ src/yolo/setup.py
